requirements.txt
yfinance
langchain-pinecone
openai
python-dotenv
langchain-community
sentence-transformers
pinecone-client
numpy
requests
scikit-learn
streamlit